+++ CHANGELOG.md
@@ -10,7 +10,11 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 ### Added
 
 -   Ability to instantiate new .NET arrays using `Array[T](dim1, dim2, ...)` syntax
+<<<<<<< HEAD
 -   Python operator method will call C# operator method for supported binary and unary operators ([#1324][p1324]).
+=======
+-   Add Interrupt method in PythonEngine
+>>>>>>> Add Interrupt method in PythonEngine
 
 ### Changed
 -   Drop support for Python 2, 3.4, and 3.5